feat: Allow hackers to delete their accounts and profile
Deleting one's account removes them completely from the system. If the
hacker is the owner of the team, their partners are also removed in the
process and everything will have to started from scratch.

In a future update, team owners will be allowed kick out members from
their team if they don't want them there. Maybe we'd add another option
for non-team owners to leave the current team they are in. But that's
not something we would really consider since membership is strictly by
invitation.