feat: Add the initial implementation to register invited hackers
This update fetches the `invite_token` that is cached when the hacker
visits the accounts creation route and registers successfully. Once they
are registered, the token (which was sent) by their team leader is used
for further processing.

This will be updated to send confirmation emails to the new hacker and
the team lead of this event.